refactor(server): isolate application grant lifecycle
